3311: Use a simple gradient instead of zigzag line
